+++ Sources/KeyboardKit/Views/Emojis/EmojiCategoryTitle.swift
@@ -15,16 +15,19 @@ import SwiftUI
 public struct EmojiCategoryTitle: View {
     
     public init(
-        title: String) {
+        title: String,
+        style: EmojiKeyboardStyle) {
         self.title = title
+        self.style = style
     }
     
     private let title: String
+    private let style: EmojiKeyboardStyle
     
     public var body: some View {
         HStack {
             Text(title)
-                .font(.footnote)
+                .font(style.categoryFont)
                 .bold()
                 .textCase(.uppercase)
                 .opacity(0.4)
@@ -37,6 +40,8 @@ public struct EmojiCategoryTitle: View {
 struct EmojiCategoryTitle_Previews: PreviewProvider {
     
     static var previews: some View {
-        EmojiCategoryTitle(title: "Hej")
+        EmojiCategoryTitle(
+            title: "Hej",
+            style: .standardPhonePortrait)
     }
 }